Create non-WId methods standalone implementation
This is needed on the way to split WId methods to their tier4/kinterprocesswindowing
framework. Initially, the analoguos non-WId methods called WId methods to avoid
code duplication. Unfortunately, that duplication has now to be done in order to
untie non-WId methods from the WId ones.

More elegant solution could be imagined but that would require the introduction of
new classes/types and the is not wanted for the moment.