Split out the KAuth management of KAction and KPushButton
It was in fact mostly duplicated code between the two classes, and
rather independent of their internals. Create a new class adding the
relevant interface using a decorator pattern. Can be extended later to
support more types.

It introduces a dependency from KAuth to QtGui though, we might want to
have two libs for KAuth (core + gui) but it seemed overkill for a single
class.