KIO listDir entries batching
KIO was not batching. At most it did:
1 signal with the first 100 entries
1 signal with the remainder

Now it's doing time based batching.
It will emit 1 signal every 300ms with the results thus far or emits as soon as all entries are in if that is before 300ms. This can be many signals if you have massive folders or a slow data source. Usually it's only 1-3 signals on the local HDD or SSD.

REVIEW: 107520