solid: report supported backends in solidConfig.cmake
This patch makes solid record which optional backends this build supports.
To be able to do this, the find_package() calls had to be moved in the top level solid
CMakeLists.txt.
It also moves the option whether to use udisk2 or udisk from the kdelibs toplevel CMakeLists.txt
into the solid toplevel CMakeLists.txt

Alex