Remove gesture factilities from KAction
The gesture facilities are now found in the KGestureMap class. Some
new methods have been added to this class to let it handle information
about each action's active and default gesture. See KDE5PORTING.html
for more information about this

I'm not sure the implementation of these methods is very efficient, as it
uses QList's searching. If that proves to slowdown shortcut editing, it
may be rewritten, but the KGestureMap class is actually not used, as one
can see in it's eventFilter method which starts with 'return false' so
this should be let as a TODO for further KF5 releases.

REVIEW: 107973