Adjust kfileplacesmodeltest to Qt5
QHash/QSet are now randomized, so turn off the randomization for the
test. It's necessary because the list of device is stored in a QSet at
some point and we need to predict the order to write the tests.

Also even without randomization the order changed compared to Qt4, so
the tests are adjusted to the new Qt5 order.

Because of the above, it is now built only for Qt5.