Rework fix for bug 279675 so that the new mimetype is used by KRun too.
When the mimetype has to be refined from the url-filename or suggested-filename,
it's not enough to show the user, we have to also pass it on to the application
and to KRun. Otherwise clicking "open with KWrite" can end up not doing anything,
due to the service->hasMimeType() check in KRun. It also didn't benefit from
the turning of shellscript into text/plain in BrowserRun::handleNonEmbeddable
since the mimetype fixup was done afterwards.

CCBUG: 279675
REVIEW: 108277