Initialise the mCrossDomain member variable in the cookies
For several months now, all my cookies would be forgotten after a kded
restart. After debugging the problem, turns out that mCrossDomain was
of value 127, which makes no sense for a boolean.

This variable has been present since 2002, which means that the
"reject cross domain cookies" feature has been broken for 10 years and
8 months.