Wait for disconnected if necessary
When close() is called a socket might not disconnect from the host
immediately. It's indicated by a temporary ClosingState after the call
to close(). When that happens wait first for the disconnect in the test
to avoid the test to randomly fail.
(cherry picked from commit 55955cdd0cf2fd15c1db4f532233a673989d9bc4)