Make plasma widgets respect plasma theme colors
Patch adds ButtonBackgroundColor to PaletteHelper. Not sure it's the best way for fixing the bugs, but it works :-)
Problem is only exposed in combination of light colorschemes and dark plasma themes which don't have a light button (so have the same button text color as for window text).
PAtch by Hrvoje Senjan

REVIEW:108525
CCMAIL:hrvoje.senjan@gmail.com