- tier2 buildsystem improvements
Add checks in the install Config.cmake files that the needed packages/imported targets (kcoreaddons
both for kauth and kconfig) are present, so we fail at cmake time if not.
This is not done for kdeqt5staging since this is temporary, so we have to live with this for some time.

Not quite sure about the other imported targets. For the imported targets for Qt it could be done too.
But what about others ? They may depend on cmake-time switches, so we can't simply hardcode them in
the Config.cmake.in file. Ok, we can configure a list of required targets to it...

Alex

CCMAIL: kde-buildsystem@kde.org, kde-frameworks-devel@kde.org