Add separate shutDown() method to QueueAPI.
The state objects (that forward the API calls to the queue
implementation) outlive the mutex, and would potentially still forward
calls the to the implementation when the mutex destructor was already
entered. shutDown() is called first, and switches the Queue into
DestructedState. The state handler returns from all API calls
immediately, effectively detaching the Queue from the Weaver object.