Remove KAction dependency on KGlobalAccelPrivate
KGlobalAccelPrivate was used only in KAction dtor to mark the shortcut
inactive when a KAction is deleted. Since we want that to work for all
QAction anyways it was needed to introduce a new mechanism.

For that we put all the known actions of KGlobalAccel in a widget, this
way the said widget will get an event when the actions are deleted, we
just spy on that.