Move fixx11h into KWindowSystem
Fixx11h is a header which is mostly needed by KWindowSystem and in
general by functionality using XLib. Because of that it makes more
sense to have the header together with other X related headers in
KWindowSystem.

At the same time the HAVE_X11 ifdef is dropped from the header. This
is not needed as the header is only intended to be used if XLib is
included. The check is moved to the user of the header which in most
has the #include already in an #ifdef HAVE_X11 block.

More information in [1].

[1]: http://lists.kde.org/?l=kde-frameworks-devel&m=136298982111644&w=2

REVIEW: 109451