Fix KDirWatch to emit dirty on parent dir when watched subdir is deleted
In Inotify mode, there's a special event for the case where a watched
dir is deleted ("IN_DELETESELF"). In the handling of that event, we must
still set the parent directory to dirty, in case another KDirWatch
only cares for the parent.

Excellent unit test by Frank Reininghaus, as usual -- thanks!

BUG: 316016
FIXED-IN: 4.10.2