Require CMake 2.8.11 RC1
The version I've changed the requirement to is not actually the RC1,
but the version build.kde.org happens to have at the moment.

 http://build.kde.org/job/cmake/lastSuccessfulBuild/

This version requirement will be bumped as new release candidates
come out, so it makes sense to run cmake master until the final
release is made available at least, as the wiki already recommends.

Followup commits will make use of the new features in this release.

CCMAIL: kde-frameworks-devel@kde.org