Shut down the global pool in a post routine to QCoreApplication.
It is not destroyed yet, but after shutDown(), all future requests into
the API will be ignored by the DestructedState handler. Since shutDown()
waits for the started threads to enter the run method if they have not
yet done so, and then controlledly makes them exit, it is ensured that
qApp still exists for every thread that starts run().