Fix border detection in Plasma dialogs
I had some problems with plasma dialogs borders, they appeared arbitrarily.
After a good couple of hours of debugging, I found out that there was some
really weird processing of the available region.

This patch removes this processing and uses QRegion itself to figure out if
the border has to be painted.

REVIEW: 109641