Add a new searchUpdated signal to KTreeWidgetSearchLine
This signal is emitted upon search text change, after the
QTreeWidget items where filtered. KTreeWidgetSearchLine does it's
filtering using an internal timer which has an arbitrary value as
seen from it's client. It's clients have no means to know the
duration of this timer. Instead, they may simply use this signal
to get notified when this timer routine finished it's filtering.

REVIEW: 109802