khtml/xml: if the content-type is null don't override mimeType and continue parsing the responseXML
In case the server does not set the content-type, the getResponseHeader("Content-Type",dummy); function will return jsNull, which is
correct.
As null is also valid, don't override the mimeType and leave it as "text/xml", so we actually do parse the reponse.

Confirmed by
http://www.w3.org/TR/XMLHttpRequest/#document-response-entity-body
"If final MIME type is not null, text/html, text/xml, application/xml, or does not end in +xml, return null and terminate these
steps."
~~
means, if it is one of these above -> continue

REVIEW:108928