Disable KHostnameD in KDED
KHostnameD monitors the hostname by polling gethostname every 5 seconds
to adapt some bits of the environment in case the hostname changes. In
systems with only a local X server this does not matter at all. In
systems using remote X you should not change your hostname, period.
In most distributions changing your hostname no longuer breaks X access
for local applications since xhost is configured correctly. In summary
there is NO reason to keep this around.

Besides that, kdontchangethehostname (binay that is called when the
hostname is changed) does not work if your DISPLAY env is like ":0"
which is the format that can be found in all distributions nowdays
(old format was domain/[unix,tcp,...]:0).

Even fixing it, most distributions use a global Xauthority pointed by
the DM or XAUTHORITY env which points to either /run/*dm/... or in
/tmp/user.... Meaning that the most important adaptation to the new
hostname that kdontchangethehostname does won't work either.

If this is ever enabled again the polling should be removed by either
using systemd-hostnamed or inotify+/etc/hostname but polling won't
be accepted.

This commits only disables the code, if nobody complains I will remove
the full code before 4.11 is out.

CCMAIL:kde-core-devel@kde.org