kjs: Fix CanvasImageData eating the color value by doing wrong alpha premultiply
We use QImage::Format_ARGB32_Premultiplied beacuse it faster for many operations.
As we directly set the red/green/blue/alpha color value via QImage::scanLine,
we must premultiply it, to keep it in the QImage::Format_ARGB32_Premultiplied format.
But as CanvasImageData does not specify any order of setting the values or does not know what values are set,
we might end up doing a wrong premultiply.
For premultiply we MUST have the correct alpha value, if it is not set it is assumed 0 (the default),
but if we premultiply a given (for example) red value with 0, its get zero, the red color value is gone.
Gone and no way to recover it.
So for example if we set for each pixel (in this order) red=255,green=0,blue=0,alpha=255,
we get a black image instead of a red one.

To fix is, change to using ARGB32, this allows us setting the color values in any order we want.

REVIEW:109998