kjs: Don't crash in PopScope if PushScope already had an Exception and didn't Push anything
As PushScope does not Push anything on Exception case, we can not Pop.
NOTE: the Exception checking on Push/Pop-Scope is not optimal.
Such a Situation, where PushScope already has a Exception can only
occour if import is used inside the the "with"-statement. As import is
executed at the beginning of the code (as val decl) and can leave a exception.
As "import" is not (yet) part of ECMA-262 the correct behavior is unknown
when this changes we can rework the bahavior, but for now just fix the crash.

NOTE2: Using import can cause problems for other statements too, but should not lead to a crash

REVIEW:110027
BUG:311512
BUG:311513