Port away from KMenu
In the case of XMLGUI it was a bit more involving as it's the main user
for context menus in menus. Rerouted the features in KMenuMenuHandler
using an event filter. Even looks better than the original version to
me, only one temporary QMenu instance is involved for the context menu
instead of creating one per QMenu in the application.