Fix crash due to uninitialized member in the containment case.
Plasma::Applet initialized the (uic-generated) publishCheckbox
only in showConfigurationInterface(), but this method isn't run
for our desktop containments; showing the config dialog is han-
dled through a signal connected to a slot in the shell implemen-
tation instead. This caused Applet::configDialogFinished() to
crash when checking for the checkbox.

This was unearthed by 94e1d5c; previously configDialogFinished()
never reached the problematic check because an earlier config
dialog id comparison would fail.

BUG:318806
REVIEW:110158