KMessageWidget: Hide decoration icon by default, add an "icon" property
This avoids confusion between the decoration icon and the close button,
especially when type is KMessageWidget::Error. This happens for example
with Dolphin when it fails to connect to a remote host.

This change also has the nice side-effect of leaving more space for the
widget text by default, but makes it possible for developers to provide
more meaningful icons when necessary.

BUG: 304775
REVIEW: 110327
FIXED-IN: 4.11.0