improve drop responsiveness ... a bit
QGraphicsView's drop method is a little broken; if you have a drag
and then it drags over another item that does not accept the drop
then your item gets a drag leave and then a drag enter immediately

really, it ought to only get a drag leave if the item the drag
passes over accepts the drag.

this makes proper state handling of drops in the panel where it
wants to only drop (and show a drag indicator) when a widget has not
accepted the drag pretty much impossible to get perfect.

appears some events also still just get dropped on the floor too
so it is possible to leave it in a state where the drag indicator
is shown.

i have l ittle hope of making this better for 4.x. need to confirm
that the drag support in QML scene graph is sane.