KConfig: properly implement revertToDefault().
Now it really removes the entry from the local file, instead of copying
the value from the global file ("defaults" file, in internal terminology).
This implements a long-standing TODO, which is now removed from the TODO file.

And it doesn't mark the KConfig as dirty if there's no local entry,
which makes closing kmail composer windows a LOT faster.

REVIEW: 103168