Figure out how the KSwitchLanguageDialog specifies the startup language
What we did was to port the code in the dialog to read the values from
QLocale, relying on KLocalizedStrings' hook to initialize QLocale properly.
So we added that hook that reads the configuration that knows what language
should be used to override the settings that Qt will use, which is the
LANGUAGE env var.

NOTE: we removed the code that, after saying that the application should be
restarted, tries to change the language on the fly. It didn't work well
(the new things were translated, but not the old things).

For more information see the review request, it has quite some literature.

REVIEW: 111178