Fix/workaround for legacy encoded filename handling
This patch works around the problem of filenames that are not valid
UTF8 strings: in KLocalePrivate::initFileNameEncoding() KDE sets the
QFile's encoding/decoding function, to to/fromUTF8() in QString, which
in turn calls QUtf8's converter function. I replaced this with the said
functions' copy/pasted version and changed it, so when it encounters an
invalid UTF8 string, it will encode it byte by byte, mapping the lower
128 their normal unicode place and the upper 128 to U+10FE00-U+10FE7F,
and the decoder reverses it.

BUG: 159241
REVIEW: 110043