KParts::StatusBarExtension: Remember state of GUIActivateEvent
When the StatusBarExtension receives a GUIActivateEvent, it shows or hides its
statusbar items accordingly, but it does not remember it: if an item is added
after receiving a GUIActivateEvent(false), the item will be shown.  Add a bool
m_activated field to keep track of the state and take it into account when a new
item is added.

REVIEW: 111345