Repair unittest.
Interesting: another side effect of porting time_t to QDateTime is that
it now includes milliseconds. So what used to be a comparison with a
precision of the second, is now precise to the millisecond (which meant
this test didn't wait for 1s as it used to).

CCMAIL: davidgiloliva@gmail.com