Move KGlobalSettings::createApplicationPalette to KColorScheme
I decided to move it to KColorScheme. We have 2 kind of users for
this method:
- applications that want to do fancy stuff with KDE and integrate
better which will have to depend on KWidgets
- and the rest, who will be able to rely on a properly set QPalette
(by whichever QPlatformTheme we're using on).

Having moved the method to KColorScheme, seems to me we can use it
from our Platform Theme, because it was doing mostly the same.

REVIEW: 111405