Resolve the KTcpSocket + KSslCertificateManager situation.
It's needed inside KIO, and by some code outside kdelibs (e.g. KImap),
but we don't want it as public API in KIO because the longer term plan
is to have this functionality in Qt. So for now, we have a copy in KIO
internally (with renamed files and namespaced classes), and another one
in kde4attic (can't be kde4support yet, since there's no replacement
available).
Once Qt has the necessary code (= storing the list of invalid certificates
approved by the user, and having the necessary hooks/signals for our dialogs
to be shown), we can kill the KIO internal copy, and move the kde4attic one
to kde4support.

CCMAIL: rich@kde.org