Add basic support for clang in FindKDE4Internal.cmake.
This fixes multiple issues when building KDE applications with clang,
such as:

- debugfull is not supported at all, it's as if one gives an 'invalid'
 build type to cmake (i.e. neither -O nor -g is present)
 - most of the "pedantic" flags are not set properly, such as those
that ensure proper visibility macros etc. pp. - none of the above is
communicated to the developer, the might compile for him but not work
as expected or he will introduce issues for others (like the windows
guys).

This patch lets clang use the same codepaths as gcc which works due
to its CLI compatibility interface.

REVIEW: 111612