Fix not found sonnet(core|ui)_export.h
0d5e8a432b7b63357696dbf2fbed5633919bff6b adjusted the include path for
the export file to be sonnet/sonnetfoo_export.h in the public header
files. This broke the build as those files cannot be found in the
current build directory.

To solve this problem the export files are now generated in a
sub-directory sonnet, so that the path matches. In addition the includes
in the private headers are adjusted to also look for sonnet/ so that
they can find it.

Patch by Martin Gräßlin

REVIEW: 111678
CCMAIL: arichardson.kde@googlemail.com