Add JobSignalsDeliveryTest.
At the time of writing, the test potentially locks up the Qt event
queue. The signal parameters carry shared pointers to qobjects that may
the last reference to the object, causing the object to be deleted in
the event loop code. In the process, QObject code tries to lock
signalSlotMutex/postEventList.mutex twice.