Get KGlobalSettings tests to pass again
They could now fail if there was more processes running with a link to
KGlobalSettings (like kded5 for some reason). In that case the
"fontsChanged" signal would be sent by all the processes each time
"refreshFonts" was emitted once. Which is really unnecessary and could
saturate the bus for nothing and create a race condition between the
font and settings tests in kglobalsettingstest.