NepomukMainModel: backport all changes from nepomuk-core
This makes all the Nepomuk1 code functional again. Because of the
removal of the local socket model, this code has stopped working
completely as it tried to connect to a local socket that no longer
existed.

Copied the code from nepomuk-core/libnepomukcore/resource/nepomukmainmodel
and changed Nepomuk2 to Nepomuk

There is still some KIO code using Nepomuk1 and that should continue to
work.