KFileMetaData (and friends) to kde4support
As far as I've understood, we should have an alternative by Nepomuk for
file previewing for KF5, this patch moves the KFileMetaInfo and the files
that depend on it to KDE4Support.

It's worth noting that there are 2 "plugins" (they're actually not plugins)
of the KPropertiesDialog that have been disabled because they had to be
moved with KFileMetaInfo. That is the kmetaprops.cpp and kpreviewprops.cpp.

REVIEW: 111897