Don't connect to KIconLoader immediately
We move the instanciation of KIconLoader in a slot so that it's only
done after the event loop is ready. Otherwise we end up using KConfig
(hidden behind KIconLoader::global()) which is a problem for later
KConfig uses.
We're limited before the event loop to reading only kdeglobals using
KConfig but KIconLoader::init needs more hence why it needs to wait (it
works in that case because none of the reading is necessary for the
plugin interface it's internal behavior).

CCMAIL: afiestas@kde.org