kshareddatacache: Add stricter per-lock sanity checks.
Instead of using complicated logic to re-map a "cache that was enlarged
under us", just treat it as a corrupt cache since cache size is
essentially invariant under the current code. This allows us to remove
the QMutex protecting just the Private d-pointer in this one case as
well.

Moving the checks made after the test for lock success also gives the
opportunity to run some quick sanity checks when the cache is locked
instead of just once when the cache is initially mapped. So we
sanity-check the eviction policy and cache version as well.