Port KWindowEffects::highlightWindows to XCB
Unit test added. Although functionality is quite similar to present
windows group, they do not share code as highlight windows deletes
the property if the list is empty. This requires a little bit
different xcb control sequence as the atom needs to be fetched before
the various checks can be done.