krecentdocument: Ensure full path is saved with entries.
This was accidentally broken in commit 48b78ce3e, which resulted in only
the filename (not the full path) being saved with KRecentDocument
entries.

The suggested fix in the reported bug is not adequate as 'dStr' can be
used later in the function, so it needs to retain the semantics it had
prior to commit 48b78ce3e. We do appreciate the prompt report though!.

BUG:323747
FIXED-IN:4.11.1