Implement Ctrl+Enter dialog acceptation in KStyle::polish()
Make it possible to make sure that we keep reacting properly to
ctrl+enter after switching from QDialog to KDialog by intercepting
QDialogButtonBox creation and injecting the desired shortcut to it.
Additionally add some porting documentation that suggests to make
the shortcut explicit and add a test to make sure that the style is
implementing it properly.

REVIEW: 112175