Move the clang compiler definitions to a separate block.
Sharing compiler settings between GCC and clang does not always work: there
are flags (such as "-fno-check-new" or "-fno-reorder-blocks") that are
specific to GCC, and nothing stops these incompatibilities from becoming
bigger in the future.

Conversely, a separate clang block allows us to pass some additional flags
to clang that would have required yet another if() in the GCC block:

 o -fdelayed-template-parsing, needed for clang not to complain when we
   build with exceptions disabled and include code that throws exceptions in
   template code we do not instantiate.

 o -Wno-return-type-c-linkage, needed at least in kdepim because one of its
   plugins exports a function with C linkage that returns a QString.

Additionally, the current checks for whether to pass -fvisibility=hidden do
not work for clang, as they parse a GCC version number that's larger than
3.x that clang returns.

Last but not least, checks for bad GCC allocators or support for some flags
which are always present in clang can be avoided altogether when we know the
compiler we are using.

(This is sort of a backport from extra-cmake-modules commit ecb29b6)

REVIEW:   112137
BUG:      323579
CCBUG:    300429 313763
FIXED-IN: 4.11.1