Always emit the claimedOwnership() signal if we have the selection
There was a logical error introduced during the porting to XCB. There
is no requirement for the old owner to destroy its selection window.
Which means we should not emit the failed signal if the window has not
been destroyed in time.

This is especially important for the situation that another Selection
Owner of the same X client takes over as the SelectionClear event is
not delivered in that case and thus the destroy never happens and the
new owner thinking it failed.

The timer is in fact only relevant for the case that the kill option
is set. We want to give the old owner the time to clean up itself
instead of forcefully killing (e.g. window manager --replace).