Check that KdePlatformTheme returns the right palette
So I have created a colorScheme with everything hardcoded to 0,128,0 and
disabled any effects so the checking gets simpler.
Regardless of this we are testing a bit of KColorScheme specially in the
colors since some effects are applied.

I can't figure out a way of chekcing that the palette is correct without
checking that KColorScheme did its job.