fillApplicationLanguages: Do not add the default locale multiple times
If QLocale cannot find the appropriate language, it then defaults to the
default locale. When adding all possible languages it is possible that
the language is not found and therefore the default lang is added
multiple times.

For me, "US English" gets added over 50 times.

We therefore, swap out the default language with C and then swap it back

REVIEW: 112288