Fix reordering of apps for a derived mimetype.
This fixes a bug introduced by 871cccc8a88a60 that made it impossible to re-order
file type associations both in System settings and in the open with list. Hence
it contains a new way of detecting duplicate (inherited) mimetype entries, that
the original was supposed to fix.

Unittest is in kde-runtime, commit 967979f (if I can commit it fast enough ;)

Patch by Mathias Tillman.
REVIEW: 111951
BUG: 321706
FIXED-IN: 4.11.1