Set 2 QApplication settings that do not have platformTheme counterpart
WheelScrollLines and ShowIconsInMenuItems are two settings that we have
in KDE, that can be set to a QApplication but that they have not been
exported as QPlatformTheme::ThemeHint, possibly because they don't make
sense in QGuiApplication.

Anyway, at the moment we are just setting them directly by using
QApplication::setWheelScrollLines (only when running on a QApplication)
and QCoreApplication::setAttribute.

Both settings are still being used in at least one time within Qt:
WheelScrollLines:
	qabstractslider.cpp

ShowIconsInMenuItems:
	QAction