Make sure the QCoreApplication instance is still around
Fixes the crashes in kdirmodeltest and kdiroperatortest introduced
by the latest merge from master. When the executable is shutting down
some DirItems are deleted during the global static dtor, in which case
the QCoreApplication instance might not be around anymore resulting in
an assert in the dbus connection pool.

This is a better fix than f37b25a696175246ac9807a9277cb1aea7af11c2 since
the assert in dbus connection pool has been introduced.