kpty: Do not require ping(1) for kpty testsuite.
Instead we simulate the effect of ping with a very small shellscript
passed directly to /bin/sh.

I would have used the bug reporter's suggested patch to check for
ping(1) during the CMake configuration process, but I wouldn't want to
stop the whole build just because ping(1) is unavailable (so I'd have to
rework the patch to make it optional). But more importantly, my distro's
ping(1) is setuid and executable but not readable, and CMake will not
find ping in this situation (see bug 10468 on Kitware's CMake
bugtracker). Apparently at least some other distros/UNIXes install ping
in this fashion as well.

Needless to say, all kpty tests still pass.

I don't intend to backport as it's technically new code, but it may be
possible to at least add a CMake comment that ping is required, post an
email to kde-packager, or something similar.

BUG:324029
FIXED-IN:4.12