LinuEditUrlDropEventFilter implemented
The replace content on URL drop functionality that was present in
KLineEdit has been moved to a separate class, that can be installed
as an event filter in any QLineEdit (KLineEdit, QComboBox,
KUrlRequester...).

KLineEdit now uses this new class when the property "urlDropsEnabled"
is enabled, but this property has been marked as deprecated in favour
of the new alternative.

REVIEW: 112177