Fix sometimes failing KArchiveTest
One of the KArchive tests is to compare if the time of creating the
archive file is the same as the time of a file inside of the archive.
This can occasionally fail thanks to some msecs delay between the
operations followed by rounding to seconds, so the times that are being
compared can end up as "18:31:00" and "18:31:01", which then fails the
test.

What I did is to take the time part out of the comparison and do a
separate comparison of the times with allowed 1 sec difference.

REVIEW: 112353