Native event filter porting for KModifierKeyInfoProvider
Ported to native event filter and to xcb-xkb by reimplementing the
events. Most parts are kept on xlib though as we don't have xkb.h
to get proper defines.

The xcb-xkb library is still considered experimental and is not
included in the xcb packages by e.g. Debian and Ubuntu. By that it
is not possible to depend on it and that's why the events are
manually specified.

REVIEW: 112443