Store servicetypes in KPluginInfo, add accessor
This way, we can prevent
KPluginInfo.service()->property("X-KDE-ServiceTypes"), and instead use
KPluginInfo.serviceTypes().

Why? The QVariantList-based ctor doesn't produce a valid d->service, as
that is read from the desktop file (or the kservice pointer passed in,
depending on the ctor used).

Two ways to get around it:

(0) Make KService understand metadata
KService being heavily based on KDesktopFile would need quite some
surgery to understand a QVariantList. I've tried constructing a
KDesktopFile from the metadata, but failed in a whole lot of sanity
checks that are run on the file. We don't have a location on disk it
understands, for example. I didn't get it to actually return me the
service types.  It also really feels like bending this class into
something it just is not.

(1) Do not use KService, make KPluginInfo suffice
This seems easier as far as matching the service type goes at least. It
also means a likely API change how to load the plugins, as we have to
change at least the return type to from KService::List to
KPluginInfo::List. Another issue is the parsing of constraints.

So let's see how far we can get with direction (1).

CCMAIL:faure@kde.org