Replace uses of NetAccess::mostLocalUrl with KIO::mostLocalUrl
Apparently there is already a job to do mostLocalUrl but it is slower
than NetAccess since it has to return a job even though some times it
already knows the answer (in case that the given url is already local).

The places where this patch reaplces NetAccess with KIO are not
performant critical so it should be ok (mostly slots that are only
called once).