Store a QVariantMap of most of the arguments
The idea here is to make KPluginInfo more flexible, so it can replace
KService in a bunch of cases, especially in the constraint parser.

* Define types for Ptr and PtrList, bringing it closer to KService API,
  which generally uses pointer-based lists.
* Remove their individual data members in the pimpl and switch users to
  the internal QVariantMap

This means:
- We can query values using the property() API, which is otherwise
  bound to an existing service, which we don't have when we
  instantiate the plugininfo from a plugin's metadata.
- KPluginInfo API is closer to KService, which is needed to make the
  constraint parser work on the one hand, and port code to

This is quite an intrusive change, and just the first one. My thinking
is that we will probably need to replace most of KServiceTypeTrader and
KService with a new KPluginTrader, using KPluginInfo::PtrList instead of
KService. Except for s/KService::List/KPluginInfo::PtrList/, this could
then be made semantically very similar.

This change itself is source-compatible. A fair bit of its churn is
because we now share the keys.

CCMAIL:dfaure@kde.org