Improve signal emitting in KSelectionOwner
I kind of broke it when adjusting to the unit test. We have to wait
for the previous owner to destroy its window before emitting the
claim signal. If we time out without the window closed, we should
emit the failed signal though. Otherwise a client will wait forever.

REVIEW: 112537