Speed up local directory listings with many files, and reduce memory usage.
Benchmarks show a speed improvement from 20ms to 14ms (for 1000 files)
and memory usage going from 92 MB to 50 MB at the peak (for 500K files).
Mark Gaiser confirms that receiving the entries in batchs doesn't slow down
the dolphin view either.

Additionally it circumvents the "expected failure due to jumbo packet size",
but that Qt5 QAbstractSocket bug is in the process of being fixed too.