Don't inspect XEvent queue in KUserTimestamp::updateUserTimestamp
updateUserTimestamp() used to fetch the current time by creating a
property notify event and wait till it is received by inspecting the
XEvent queue. That doesn't work with XCB anymore, but Qt provides
this functionality for us, so just use that.

At the same time the argument is changed from int to unsigned long
as that's what the method in QX11Info returns.

REVIEW: 112684